Design and Implement Language Selection Component Fixes #3. Completed
