+++ src/app/settings/settings.component.html
@@ -13,9 +13,13 @@
       <mat-expansion-panel-header>
         <mat-panel-title>
           <mat-icon>language</mat-icon>
+          Language:
         </mat-panel-title>
         <mat-panel-description>
-          {{langOrigin}}<span width = "5px"></span>{{langEn}}       
+          <span width = "5px"></span>
+          {{langOrigin}}
+          <span width = "5px"></span>
+          {{langEn}}       
         </mat-panel-description>
       </mat-expansion-panel-header>
       <app-language-selection></app-language-selection>
@@ -24,7 +28,7 @@
     <mat-expansion-panel>
       <mat-expansion-panel-header>
         <mat-panel-title>
-          <mat-icon>style</mat-icon>
+          <mat-icon>palette</mat-icon>
           Appearence
         </mat-panel-title>
         <mat-panel-description>
