+++ src/app/settings/settings.component.ts
@@ -1,4 +1,4 @@
-import {Component, ViewChild} from '@angular/core';
+import {Component, ViewChild, OnInit} from '@angular/core';
 import {MatAccordion, MatExpansionModule} from '@angular/material/expansion';
 import {MatDatepickerModule} from '@angular/material/datepicker';
 import {MatInputModule} from '@angular/material/input';
@@ -8,9 +8,9 @@ import {MatButtonModule} from '@angular/material/button';
 import {MatNativeDateModule} from '@angular/material/core';
 import {MatCardModule} from '@angular/material/card';
 import {MatProgressBarModule} from '@angular/material/progress-bar';
-import {LanguageDescription, LanguageSelectionComponent} from '../language-selection/language-selection.component'
-import { LanguageSelectionNotificationService } from '../language-selection/language-selection-notification.service';
-import { Subscription } from 'rxjs/internal/Subscription';
+import {LanguageSelectionComponent} from '../language-selection/language-selection.component'
+import {LanguageDescription, LanguageSelectionNotificationService,  inSupportedLanguages} from '../language-selection/language-selection-notification.service';
+import {Subscription} from 'rxjs/internal/Subscription';
 
 
 /**
@@ -34,13 +34,14 @@ import { Subscription } from 'rxjs/internal/Subscription';
   templateUrl: './settings.component.html',
   styleUrl: './settings.component.scss'
 })
-export class SettingsComponent {
+export class SettingsComponent implements OnInit  {
   @ViewChild(MatAccordion) accordion?: MatAccordion;
 
   private subscription: Subscription;
 
   langOrigin: string = ""
   langEn: string = ""
+  langEtfTag = "" 
 
   constructor(private languageSelectionNotificationService: LanguageSelectionNotificationService) {
     this.subscription = this
@@ -48,6 +49,27 @@ export class SettingsComponent {
       .subscribe((selectedLanguage: LanguageDescription) => {
       this.langOrigin = selectedLanguage.originalName;
       this.langEn = selectedLanguage.enName
+      this.langEtfTag = selectedLanguage.ietfTag
     });
+  }
+
+  ngOnInit() {
+    this.trySetLanguage();
+  }
+
+   trySetLanguage() {
+
+    let savedLangEtfTag = localStorage.getItem("langEtfTag")
+
+    if(typeof savedLangEtfTag !== 'string'){
+      savedLangEtfTag = navigator.language;
+      console.log('Browser Language:', savedLangEtfTag);
+    }
+
+    if(!inSupportedLanguages(savedLangEtfTag)){
+      savedLangEtfTag = "en-US";
+    }
+
+    this.languageSelectionNotificationService.setLanguage(savedLangEtfTag as string)
   }
 }
